customer-service/app/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-environ>=0.11
psycopg2-binary>=2.9
gunicorn>=21.2
requests>=2.31
